feat: Update Drush commands in entrypoint.sh
The entrypoint.sh file was updated to improve the execution of Drush commands. The `drush cr` command was moved to the end of the script to clear the cache after running other commands. Additionally, the `drush updb -y` command was added before the `drush cr` command to ensure that any necessary database updates are applied before clearing the cache.